requirements.txt
flask
keras.models
tensorflow.keras.preprocessing.sequence
numpy
pickle
time
os
gunicorn
kagglehub